requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-chroma
langchain-huggingface
sentence-transformers
sqlglot
python-dotenv
pydantic